+++ gradle.properties
@@ -94,7 +94,7 @@ ANDROIDX_APPCOMPAT_VERSION=1.4.1
 ANDROIDX_CAMERA_VERSION=1.1.0-rc01
 ANDROIDX_COMPOSE_COMPILER_VERSION=1.2.0-beta03
 ANDROIDX_COMPOSE_MATERIAL3_VERSION=1.0.0-alpha13
-ANDROIDX_COMPOSE_VERSION=1.1.1
+ANDROIDX_COMPOSE_VERSION=1.2.0-beta03
 ANDROIDX_CONSTRAINTLAYOUT_VERSION=1.0.0
 ANDROIDX_CORE_VERSION=1.7.0
 ANDROIDX_ESPRESSO_VERSION=3.5.0-alpha07
