+++ ui-integration-test-lib/src/androidTest/java/co/electriccoin/zcash/ui/integration/test/common/IntegrationTestingActivity.kt
@@ -1,0 +1,5 @@
+package co.electriccoin.zcash.ui.integration.test.common
+
+import androidx.activity.ComponentActivity
+
+class IntegrationTestingActivity : ComponentActivity()
