+++ ui-integration-test-lib/src/androidTest/java/co/electriccoin/zcash/ui/integration/test/screen/scan/view/ScanViewTest.kt
@@ -11,9 +11,9 @@ import androidx.compose.ui.test.performClick
 import androidx.test.filters.LargeTest
 import co.electriccoin.zcash.test.UiTestPrerequisites
 import co.electriccoin.zcash.ui.R
+import co.electriccoin.zcash.ui.integration.test.common.IntegrationTestingActivity
 import co.electriccoin.zcash.ui.integration.test.getPermissionPositiveButtonUiObject
 import co.electriccoin.zcash.ui.integration.test.getStringResource
-import co.electriccoin.zcash.ui.integration.test.screen.scan.TestScanActivity
 import co.electriccoin.zcash.ui.integration.test.waitForDeviceIdle
 import co.electriccoin.zcash.ui.screen.scan.ScanTag
 import co.electriccoin.zcash.ui.screen.scan.model.ScanState
@@ -28,7 +28,7 @@ import kotlin.time.Duration.Companion.milliseconds
 class ScanViewTest : UiTestPrerequisites() {
 
     @get:Rule
-    val composeTestRule = createAndroidComposeRule<TestScanActivity>()
+    val composeTestRule = createAndroidComposeRule<IntegrationTestingActivity>()
 
     private lateinit var testSetup: ScanViewTestSetup
