+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/common/UiTestingActivity.kt
@@ -1,0 +1,5 @@
+package co.electriccoin.zcash.ui.common
+
+import androidx.activity.ComponentActivity
+
+class UiTestingActivity : ComponentActivity()
