+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/screen/onboarding/integration/OnboardingIntegrationTest.kt
@@ -1,14 +1,14 @@
-package co.electriccoin.zcash.ui.screen.onboarding.view
+package co.electriccoin.zcash.ui.screen.onboarding.integration
 
 import androidx.compose.ui.test.junit4.StateRestorationTester
-import androidx.compose.ui.test.junit4.createAndroidComposeRule
+import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.test.filters.MediumTest
 import co.electriccoin.zcash.test.UiTestPrerequisites
 import co.electriccoin.zcash.ui.R
-import co.electriccoin.zcash.ui.screen.onboarding.TestOnboardingActivity
 import co.electriccoin.zcash.ui.screen.onboarding.model.OnboardingStage
+import co.electriccoin.zcash.ui.screen.onboarding.view.OnboardingTestSetup
 import co.electriccoin.zcash.ui.test.getStringResource
 import org.junit.Assert.assertEquals
 import org.junit.Rule
@@ -17,7 +17,7 @@ import org.junit.Test
 // TODO [#382]: https://github.com/zcash/secant-android-wallet/issues/382
 class OnboardingIntegrationTest : UiTestPrerequisites() {
     @get:Rule
-    val composeTestRule = createAndroidComposeRule<TestOnboardingActivity>()
+    val composeTestRule = createComposeRule()
 
     private fun newTestSetup(initialStage: OnboardingStage) = OnboardingTestSetup(composeTestRule, initialStage)
 
@@ -48,26 +48,5 @@ class OnboardingIntegrationTest : UiTestPrerequisites() {
         restorationTester.emulateSavedInstanceStateRestore()
 
         assertEquals(OnboardingStage.More, testSetup.getOnboardingStage())
-    }
-
-    @Test
-    @MediumTest
-    fun current_stage_restoration_activity() {
-        val testSetup = newTestSetup(OnboardingStage.ShieldedByDefault)
-        testSetup.setDefaultContent()
-
-        assertEquals(OnboardingStage.ShieldedByDefault, testSetup.getOnboardingStage())
-
-        composeTestRule.onNodeWithText(getStringResource(R.string.onboarding_next)).also {
-            it.performClick()
-        }
-
-        assertEquals(OnboardingStage.UnifiedAddresses, testSetup.getOnboardingStage())
-
-        composeTestRule.activityRule.scenario.onActivity {
-            it.recreate()
-        }
-
-        assertEquals(OnboardingStage.UnifiedAddresses, testSetup.getOnboardingStage())
     }
 }
