--- ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/screen/update/TestUpdateActivity.kt
@@ -1,36 +1,0 @@
-package co.electriccoin.zcash.ui.screen.update
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.ui.Modifier
-import co.electriccoin.zcash.ui.design.component.GradientSurface
-import co.electriccoin.zcash.ui.design.theme.ZcashTheme
-import co.electriccoin.zcash.ui.screen.update.fixture.UpdateInfoFixture
-
-class TestUpdateActivity : ComponentActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setupUiContent()
-    }
-
-    private fun setupUiContent() {
-        setContent {
-            ZcashTheme {
-                GradientSurface(
-                    Modifier
-                        .fillMaxWidth()
-                        .fillMaxHeight()
-                ) {
-                    WrapUpdate(
-                        this,
-                        UpdateInfoFixture.new(appUpdateInfo = null)
-                    )
-                }
-            }
-        }
-    }
-}
